+++ chirho/indexed/internals.py
@@ -1,4 +1,3 @@
-import functools
 import numbers
 import typing
 from typing import Dict, Hashable, Optional, TypeVar, Union
@@ -8,6 +7,7 @@ import pyro.distributions
 import pyro.infer.reparam
 import torch
 from pyro.poutine.indep_messenger import CondIndepStackFrame, IndepMessenger
+from typing_extensions import ParamSpec
 
 from chirho.indexed.ops import (
     IndexSet,
@@ -19,6 +19,7 @@ from chirho.indexed.ops import (
     union,
 )
 
+P = ParamSpec("P")
 K = TypeVar("K")
 T = TypeVar("T")
 
@@ -280,27 +281,3 @@ def get_sample_msg_device(
 @pyro.poutine.runtime.effectful(type="add_indices")
 def add_indices(indexset: IndexSet) -> IndexSet:
     return indexset
-
-
-if not typing.TYPE_CHECKING:
-    if tuple(map(int, pyro.__version__.split(".")[:2])) >= (1, 9):
-
-        # This is a temporary patch to work around a bug introduced in Pyro 1.9.0.
-        # TODO this should be removed once the bug is fixed upstream.
-
-        class _PatchDefaultConstraintKwarg(pyro.poutine.messenger.Messenger):
-            @staticmethod
-            def _pyro_param(msg):
-                # pyro.param's constraint argument has a default value of "real",
-                # but Pyro 1.9.0 includes dummy param statements in pyro.nn.PyroModule
-                # that are incorrectly missing a value for "constraint" (default or otherwise).
-                # This handler adds the default value ahead of any trace handlers that might
-                # otherwise contain incorrectly formed "param" nodes.
-                if "constraint" not in msg["kwargs"]:
-                    msg["kwargs"]["constraint"] = pyro.distributions.constraints.real
-
-        # Currently, the only known place where this bug affects chirho is in model rendering,
-        # so we only patch the function pyro.infer.inspect.get_model_relations used in rendering.
-        pyro.infer.inspect.get_model_relations = functools.wraps(
-            pyro.infer.inspect.get_model_relations
-        )(_PatchDefaultConstraintKwarg()(pyro.infer.inspect.get_model_relations))
