+++ chirho/indexed/ops.py
@@ -2,7 +2,6 @@ import functools
 import operator
 import typing
 from typing import (
-    Callable,
     Dict,
     Hashable,
     Iterable,
@@ -19,6 +18,8 @@ import pyro
 import pyro.poutine.indep_messenger
 import torch
 from typing_extensions import ParamSpec
+
+from .. import _pyro_patch
 
 P = ParamSpec("P")
 T = TypeVar("T")
@@ -264,22 +265,7 @@ def scatter(
     raise NotImplementedError
 
 
-def _just(fn: Callable[P, Optional[T]]) -> Callable[P, T]:
-
-    if not typing.TYPE_CHECKING:
-        return fn
-
-    @functools.wraps(fn)
-    def _wrapped(*args: P.args, **kwargs: P.kwargs) -> T:
-        result = fn(*args, **kwargs)
-        if typing.TYPE_CHECKING:
-            assert result is not None
-        return result
-
-    return _wrapped
-
-
-@_just
+@_pyro_patch._just
 @pyro.poutine.runtime.effectful(type="scatter_n")
 def scatter_n(values: Dict[IndexSet, T], *, result: Optional[T] = None, **kwargs) -> T:
     """
@@ -328,7 +314,7 @@ def cond(fst, snd: T, case: Optional[Union[bool, torch.Tensor]] = None, **kwargs
     raise NotImplementedError(f"cond not implemented for {type(fst)}")
 
 
-@_just
+@_pyro_patch._just
 @pyro.poutine.runtime.effectful(type="cond_n")
 def cond_n(values: Dict[IndexSet, T], case: Union[bool, torch.Tensor], **kwargs) -> T:
     assert len(values) > 0
@@ -346,7 +332,7 @@ def cond_n(values: Dict[IndexSet, T], case: Union[bool, torch.Tensor], **kwargs)
     return result
 
 
-@_just
+@_pyro_patch._just
 @pyro.poutine.runtime.effectful(type="get_index_plates")
 def get_index_plates() -> (
     Mapping[str, pyro.poutine.indep_messenger.CondIndepStackFrame]
